requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
asyncio==3.4.3
SQLAlchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
alembic==1.12.1
pydantic==2.5.2
structlog==23.2.0
openai==1.3.7
python-magic==0.4.27 